addres todo in comment
I leave this comment here because either
* A
We don't need to read here and this is code left from some other purpose
* B
This adresses some bug where the image must be read first before awt can
to something with the bytes. Or this is to check the image can be read
from.

Signed-off-by: MaSven <dev@mail.smarquardt.space>