Implement custom annotation processor to to ease the creation of more meaningful provider/service annotations. Currently all services need to be annotated using @AutoService(<service class>) which is often not very conceptual. Also for the sake of decoupling views from models the current ServiceLoader api doesn't provide the option to have any generic constraints.
The first problem is solved by providing the @ServiceSpec(<service class>) meta annotation,
which can be used to create new service annotations, where the annotated have to be assignable to the given <service class>.
For general single-use service declaration the @Service(<service class>) annotation replaces @AutoService(<service class>).

The second problem is solved by providing the @ProviderSpec(inputBound, outputBound) meta annotation which can
be used to create provider annotations. Classes annotated with @ProviderSpec need to implement the ProviderFor<A, B> interface,
such that A is assignable from inputBound and B is assignable to outputBound.
These types are checked when requesting the provided value. Request are fulfilled in the following manner:

**Note** that A and B are treated throughout as if type erasure had been applied.

For an object a of type A which request an object of type B a provider which is "assignable" to ProviderFor<? super A, ? extend B>.
("assignable" in quotes because java won't actually let one assign one to the other, but assignable is meant in the way that the following is legal:
````java
ProviderFor<? super A, ? extends B> provider = ...
A a = ...
B b = provider.provide(a);
````
).
If there are multiple candidates which match, then the provider with the most specific inout and the least specific output type is used.
i.e. if there are type C < B < A (where < means "is subtype of"), and P<T, K> := ProviderFor<T, K> then
- requesting P<C, A> with P<A, A>, P<B, A> and P<C, A> available will result in P<C, A> being selected.
- requesting P<A, A> with P<A, A>, P<A, B> and P<A, C> available will result in P<C, A> being selected.
If there are multiple providers which fulfill the requested specification which cannot be compared
(e.g. P<A1, B>, P<A2, B> where A1 and A2 both are subtypes of A but neither is a subtype of the other) then a warning is issued and no guarantee is given
on which provider will be used.

There also is a general @Provider annotation which imposes no bounds on the input and output type.

In contrast to AutoService the used approach is one of an isolating annotation processor which allows for better incremental compilation.

The ServiceManager is responsible for handling provider/service requests and uses the ServiceLoader internally.